Unify configuration
