+++ scripts/jest/environment.js
@@ -1,7 +1,11 @@
 /* eslint-disable */
-global.__DEV__ = true;
 
-// For testing DOM Fiber.
+const NODE_ENV = process.env.NODE_ENV;
+if (NODE_ENV !== 'development' && NODE_ENV !== 'production') {
+  throw new Error('NODE_ENV must either be set to development or production.');
+}
+global.__DEV__ = NODE_ENV === 'development';
+
 global.requestAnimationFrame = function(callback) {
   setTimeout(callback);
 };
