--- scripts/jest/prod/environment.js
@@ -1,21 +1,0 @@
-/* eslint-disable */
-global.__DEV__ = false;
-
-// For testing DOM Fiber.
-global.requestAnimationFrame = function(callback) {
-  setTimeout(callback);
-};
-
-global.requestIdleCallback = function(callback) {
-  return setTimeout(() => {
-    callback({
-      timeRemaining() {
-        return Infinity;
-      },
-    });
-  });
-};
-
-global.cancelIdleCallback = function(callbackID) {
-  clearTimeout(callbackID);
-};
